+++ internal/core/domain/entity/delivery.go
@@ -4,5 +4,5 @@ type Delivery struct {
 	ID      int     `db:"id"`
 	Address string  `db:"address"`
 	Weight  float64 `db:"weight"`
-	Cost    int     `db:"cost"`
+	Cost    float64 `db:"cost"`
 }
