+++ internal/core/domain/entity/product.go
@@ -4,8 +4,8 @@ type Product struct {
 	ID         int     `db:"id"`
 	Name       string  `db:"name"`
 	CategoryID int     `db:"category_id"`
-	Length     int     `db:"length"`
-	Price      int     `db:"price"`
+	Length     float64 `db:"length"`
+	Price      float64 `db:"price"`
 	Weight     float64 `db:"weight"`
 	Code       string  `db:"code"`
 }
